% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmr_stats.R
\name{average_moving_range}
\alias{average_moving_range}
\title{Compute the average moving range}
\usage{
average_moving_range(data, date_col, moving_range_col, ...)
}
\arguments{
\item{data}{A dataframe with at least two columns.}

\item{date_col}{The unquoted name of the column which contains the date for
each value.}

\item{moving_range_col}{The unquoted name of the column which contains the
difference between successive values, known as the moving ranges.}

\item{...}{\code{\link[rlang:args_data_masking]{<data-masking>}} Optionally,
pass expressions to filter the data, which limits the range of values over
which the mean will be computed.}
}
\value{
A numeric value.
}
\description{
Calculate the average (mean) of the moving ranges.
}
\examples{
# Compute moving ranges.
data_with_MRs <- moving_ranges(pounds_scrapped, date, value)

# Compute average moving range for values that occur prior to January 2024.
average_moving_range(data_with_MRs, date, moving_range, date < as.Date("2024-01-01"))
}
