% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descStats.R
\name{CI95}
\alias{CI95}
\title{95\% confidence interval}
\usage{
CI95(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector containing a two values: the lower and upper bounds
 of the 95\% confidence interval.
}
\description{
95\% confidence interval (ignores missing values).
}
\examples{
observations <- c(0, 1, 2, 4, 5, 6, 7, NA, 9, 10)
CI95(observations)

}
\seealso{
Other descriptive functions: 
\code{\link{CV}()},
\code{\link{L95CI}()},
\code{\link{Max}()},
\code{\link{Mean}()},
\code{\link{Median}()},
\code{\link{Min}()},
\code{\link{N_NA}()},
\code{\link{N_Obs}()},
\code{\link{N_Zero}()},
\code{\link{SD}()},
\code{\link{SE}()},
\code{\link{Sum}()},
\code{\link{U95CI}()},
\code{\link{Var}()},
\code{\link{sumFunShort}()},
\code{\link{sumFun}()}
}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
\concept{descriptive functions}
