% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminalTools.R
\name{colIdx}
\alias{colIdx}
\title{Find the numerical address for columns in a data.frame.}
\usage{
colIdx(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A vector of column names, 'named' by their numerical address.
}
\description{
Find the numerical address for columns in a data.frame.
}
\examples{

colIdx(friutSalad())

}
\seealso{
Other terminal tools: 
\code{\link{dfScan}()}
}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
\concept{terminal tools}
