% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmr_stats.R
\name{compute_average}
\alias{compute_average}
\title{Compute either the average value or the average moving range}
\usage{
compute_average(data, date_col, value_col, moving_ranges = FALSE, ...)
}
\arguments{
\item{data}{A dataframe with at least two columns.}

\item{date_col}{The unquoted name of the column which contains the date for
either each value or each moving range.}

\item{value_col}{he unquoted name of the column which contains either the
values or the difference between successive values (known as the moving
ranges).}

\item{moving_ranges}{Compute the average of the values (\code{FALSE}) or the
average of the moving ranges (\code{TRUE})? By default, the average values will
be returned.}

\item{...}{\code{\link[rlang:args_data_masking]{<data-masking>}} Optionally,
pass expressions to filter the data, which limits the range of values or
moving ranges over which the mean will be computed.}
}
\value{
A numeric value.
}
\description{
Compute either the average value or the average moving range
}
\examples{
# Compute moving ranges.
data_with_MRs <- moving_ranges(pounds_scrapped, date, value)

# Compute average moving range for values that occur prior to January 2024.
compute_average(
  data_with_MRs,
  date,
  moving_range,
  moving_ranges = TRUE,
  date < as.Date("2024-01-01")
)
}
