% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_coordRange48States.R
\docType{data}
\name{coordRange48States}
\alias{coordRange48States}
\title{Extreme Coordinates for the 48 States}
\format{
A data frame with 49 rows (48 states + D.C.) and the following 4 variables.
 \describe{
   \item{first}{a numeric vector of extreme eastern longitude values}
   \item{second}{a numeric vector of extreme western longitude values}
   \item{third}{a numeric vector of extreme northern latitude values}
   \item{fourth}{a numeric vector of extreme southern latitude values}
 }
}
\source{
unknown
}
\usage{
data(coordRange48States)
}
\description{
Somewhere I found polygons for the 48 states and extracted the minimum and maximum latitudes and longitudes for each state.  These coordinates help when you are figuring out where to crop maps.
}
\examples{
str(coordRange48States)
head(coordRange48States)

}
\keyword{datasets}
