% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_fishNames.R
\docType{data}
\name{d9Sportfish}
\alias{d9Sportfish}
\title{District 9 Sportfish}
\format{
A data frame with 36 rows of sportfish species and the following 4 variables:
 \describe{
   \item{first}{a character vector of NCWRC abbreviations}
   \item{second}{a character vector of genus and species}
   \item{third}{a character vector of common names}
   \item{fourth}{a character vector of wheather of not the species is a sportfish that is present in District-9}
 }
}
\source{
unknown
}
\usage{
data(d9Sportfish)
}
\description{
A data.frame of sportfishes in NCWRC District 9.
}
\examples{
str(d9Sportfish)
head(d9Sportfish)

}
\keyword{datasets}
