% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_fishNames.R
\docType{data}
\name{d9SportfishVector}
\alias{d9SportfishVector}
\title{District 9 Sportfish Vector}
\format{
A vector of District-9 sportfish abreviations labeled with their common names.
}
\source{
unknown
}
\usage{
data(d9SportfishVector)
}
\description{
A vector of sportfishes in NCWRC District 9.
}
\examples{
str(d9SportfishVector)
head(d9SportfishVector)

}
\keyword{datasets}
