% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTools.R
\name{ordinalDate}
\alias{ordinalDate}
\alias{daysFromToday}
\title{Assign ordinal dates.}
\usage{
ordinalDate(date)

daysFromToday(x = 0)
}
\arguments{
\item{x}{Numeric. The number of days in the future you are interested in.}

\item{X}{A date in R's native format.}
}
\value{
A vector or ordinal dates.

A vector of class date with the future date.
}
\description{
Function that converts R dates into ordinal dates. For example "2008-07-14" becomes 195. Ordinal dates can be useful for graphing. For example if you monitor some variable continually for multiple years and you want to graph them on a common January to December X-axis.

Find what day is X days from today.
}
\note{
This function returns NA for leap year days (Februay 29) and is not vectorized.  If you want to call it on multiple observations in a vector or a data.frame column use sapply().  See examples.

This is easier than manually counting days on a calendar. For example, I have to service water quality loggers every 90 days before their batteries expire. This function makes it easy to find 90 days from now.
}
\examples{
example <- as.Date("07/01/1999", format = '\%m/\%d/\%Y')
ordinalDate(example)

example2 <- as.Date(c("1/1/2000", "07/01/1999", "12/31/1970"), format = '\%m/\%d/\%Y')
sapply(example2, ordinalDate)


daysFromToday(90)

}
\seealso{
Other date tools: 
\code{\link{today}()},
\code{\link{yy2yyyy}()}

Other date tools: 
\code{\link{today}()},
\code{\link{yy2yyyy}()}
}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}

A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
\concept{date tools}
