% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmm2dd.R
\name{ddmm2dd}
\alias{ddmm2dd}
\title{Convert a GPS coordinate in degrees and minutes (ddmm.mmm) to decimal degrees (dd.ddd).}
\usage{
ddmm2dd(ddmm)
}
\arguments{
\item{ddmm}{A numeric vector of a latitude or longitude in the format: ddmm.mmmm.}
}
\value{
A numeric vector of a latitude or longitude in the format: dd.dddddd
}
\description{
Convert historically common degrees and minutes (ddmm.mmm) to more contempoary decimal degrees (dd.ddd).
}
\section{testing}{
 I haven't found a good example to test this against.
}

\examples{
## Convert ddmm.mmm to dd.dddd
coord <- c(-8355.99)
ddmm2dd(coord)

}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
