% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminalTools.R
\name{dfScan}
\alias{dfScan}
\alias{basicSysInfo}
\title{View a subset of the rows in a large data.frame.}
\usage{
dfScan(df, N = 50)

basicSysInfo(...)
}
\arguments{
\item{df}{Data.frame.}

\item{N}{Number. The number of rows you want to display. Must be >= 3.}
}
\value{
A data.frame of the sytematically subsampled rows.

Does not return an object only prints information to the console.
}
\description{
View the first, last, and evenly spaced rows throughout a data.frame.

Displays basic information on R, hardware, operating system, and date.
}
\note{
This is helpful for getting a feel for what's going on in a very large data.frame if it is too large to easially display.

This is helpful for adding system information to script outputs.
}
\examples{

dfScan(friutSalad(1000))

## Useage:
basicSysInfo()

}
\references{
A more refined version was published in the FSA package as peek().
}
\seealso{
Other terminal tools: 
\code{\link{colIdx}()}

Other terminal tools: 
\code{\link{colIdx}()}
}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}

A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
\concept{terminal tools}
