% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCoords.R
\name{findCoords}
\alias{findCoords}
\title{Find GPS coordinates of anything}
\usage{
findCoords(place = NA)
}
\arguments{
\item{x}{A string describing the location of interest.}
}
\value{
A data.frame containing the longitude and latitude.
}
\description{
Find GPS coordinates of anything.
}
\examples{
findCoords ("The White House")

}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
