% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{flow}
\alias{flow}
\title{Stream flow.}
\usage{
flow(
  station_ft = NA,
  depth_ft = NA,
  vel_ft_per_sec = NA,
  stream = "None",
  site = "None",
  date = "None"
)
}
\arguments{
\item{station_ft}{A numeric column in a data.frame. The transect station points recorded in decimal feet.}

\item{depth_ft}{A numeric column in a data.frame. The depths at the transect station points recorded in decimal feet.}

\item{vel_ft_per_sec}{A numeric column in a data.frame. The depths at the transect station points recorded in decimal feet.}

\item{stream}{Character string for the stream name. This information is only used to add metadata to the output and can be left blank. If you reference a data.frame column, then the function uses the only the value in the first row.}

\item{site}{Character string for site name. This information is only used to add metadata to the output and can be left blank. If you reference a data.frame column, then the function uses the value in the first row.}

\item{date}{Character string for date. This information is only used to add metadata to the output and can be left blank. If you reference a data.frame column, then function uses only the value in the first row.}
}
\value{
A list including the Stream Name, Site, Date, Transect Width (ft), Transect Weighted Mean Depth (ft), Transect Cross Sectional Area (sq_ft), Transect Weighted Mean Velocity (ft/s), Interval Calculations, and Discharge (cfs, cms, gps)
}
\description{
Calculate stream flow from interval velocity and depth measurements.
}
\section{testing}{
 #I checked this against Brady Dodd's Excel Sheet with Data from Peckerwood Creek, NC.
}

\examples{
## Read a CSV example file internal to wheelR package.
peckerwood <- read.csv(system.file("extdata", "peckerwood.csv", package = 'wheelR'), na='.', stringsAsFactors=FALSE)))

## In this example the following three calls produce the same results.

flow(station_ft = peckerwood$station_ft, depth_ft = peckerwood$depth_ft, vel_ft_per_sec=peckerwood$vel_ft_per_sec, stream="peckerwood", site="site", date="date")
flow(station_ft = peckerwood$station_ft, depth_ft = peckerwood$depth_ft, vel_ft_per_sec=peckerwood$vel_ft_per_sec, stream=peckerwood$stream, site=peckerwood$site, date=peckerwood$date)
flow(peckerwood$station_ft, peckerwood$depth_ft, peckerwood$vel_ft_per_sec, peckerwood$stream, peckerwood$site, peckerwood$date)

}
\references{
This uses the Midsection Method from Turnipseed, D. P., and Sauer, V. B. 2010. Discharge measurements at gaging stations. U.S. Geological Survey, Washington.
}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
