% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruitSalad.R
\name{fruitSalad}
\alias{fruitSalad}
\title{Simple data.frame for testing functions and unfamiliar commands.}
\usage{
fruitSalad(N = 20)
}
\arguments{
\item{N}{Numeric vector with length = 1 to describe the number of rows in the output data.frame. Defaults to 20.}
}
\value{
A data.frame with information about fictitious salads.
}
\description{
Creates a simple data.frame with a variety of variable types for testing purposes.
}
\examples{
fruitSalad()
fruitSalad(50)

}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
