% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{ha2ac}
\alias{ha2ac}
\title{Convert hectares to acres}
\usage{
ha2ac(x)
}
\arguments{
\item{x}{A numeric vector of measurements in hectares.}
}
\value{
A numeric vector of measurements in acres.
}
\description{
Convert hectares to acres.
}
\examples{
ha2ac <- c(0, 1, 2, 4, 5, 6, 7, NA, 9, 10)

}
\references{
The converion factor is from www.onlineconversion.com.
}
\seealso{
Other unit conversion functions: 
\code{\link{ac2ha}()},
\code{\link{c2f}()},
\code{\link{cubft2cubm}()},
\code{\link{cubft2gal}()},
\code{\link{cubm2cubft}()},
\code{\link{deg2rad}()},
\code{\link{f2c}()},
\code{\link{ft2m}()},
\code{\link{g2lb}()},
\code{\link{gal2cubft}()},
\code{\link{in2mm}()},
\code{\link{kg2lb}()},
\code{\link{km2mi}()},
\code{\link{lb2g}()},
\code{\link{lb2kg}()},
\code{\link{m2ft}()},
\code{\link{mi2km}()},
\code{\link{mm2in}()},
\code{\link{rad2deg}()}
}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
\concept{unit conversion functions}
