% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{in2mm}
\alias{in2mm}
\title{Convert inches to millimeters}
\usage{
in2mm(x)
}
\arguments{
\item{x}{A numeric vector of measurements in inches.}
}
\value{
A numeric vector of measurements in millimeters.
}
\description{
Convert inches to millimeters.
}
\examples{
in2mm <- c(0, 1, 2, 4, 5, 6, 7, NA, 9, 10)

}
\references{
The converion factor is from www.onlineconversion.com.
}
\seealso{
Other unit conversion functions: 
\code{\link{ac2ha}()},
\code{\link{c2f}()},
\code{\link{cubft2cubm}()},
\code{\link{cubft2gal}()},
\code{\link{cubm2cubft}()},
\code{\link{deg2rad}()},
\code{\link{f2c}()},
\code{\link{ft2m}()},
\code{\link{g2lb}()},
\code{\link{gal2cubft}()},
\code{\link{ha2ac}()},
\code{\link{kg2lb}()},
\code{\link{km2mi}()},
\code{\link{lb2g}()},
\code{\link{lb2kg}()},
\code{\link{m2ft}()},
\code{\link{mi2km}()},
\code{\link{mm2in}()},
\code{\link{rad2deg}()}
}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
\concept{unit conversion functions}
