% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmr_stats.R
\name{lower_natural_process_limit}
\alias{lower_natural_process_limit}
\title{Compute the Lower Natural Process Limit}
\usage{
lower_natural_process_limit(avg_moving_range, avg_value)
}
\arguments{
\item{avg_moving_range}{Numeric value, expressing the average of the
moving ranges.}

\item{avg_value}{Numeric value, expressing the average of the values.}
}
\value{
Numeric value.
}
\description{
Compute the Lower Natural Process Limit
}
