% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeTools.R
\name{minsec2sec}
\alias{minsec2sec}
\title{Convert a time measurement in MM:SS format to seconds.}
\usage{
minsec2sec(minsec)
}
\arguments{
\item{minsec}{A character vector with the time measurement as MM:SS. For example 12 mins and 30 seconds would be "12:30".}
}
\value{
A vector of the time measurement in seconds.
}
\description{
Converts a time measurement in MM:SS format to seconds.
}
\note{
This function is not vectorized. Please use sapply() to call it on multiple values.  sapply() will return a named vector, if you prefer an un-named vector use unname() in the call. See examples.
  Don't do anything crazy because I haven't completely armored the function against stupidity.  For example the following will break the function: ":30", "30:", "AB:CD"
}
\examples{

minsec2sec('12:30')

testData <- c("01:30", "1:30", "100:00")

sapply(testData, min.sec2sec) #Returns a named vector

sapply(unname(testData, min.sec2sec)) #Returns an unnamed vector

}
\seealso{
Other time tools: 
\code{\link{now}()}
}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
\concept{time tools}
