% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmr_stats.R
\name{moving_ranges}
\alias{moving_ranges}
\title{Compute the moving ranges}
\usage{
moving_ranges(data, date_col, value_col)
}
\arguments{
\item{data}{A dataframe with at least two columns.}

\item{date_col}{The unquoted name of the column which contains the date for
each value.}

\item{value_col}{The unquoted name of the column which contains the values
for which the moving ranges will be computed.}
}
\value{
A dataframe with one column greater than the number in \code{data}.
}
\description{
Find the difference between successive values, known as the
moving ranges.
}
\examples{
moving_ranges(pounds_scrapped, date, value)
}
