% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descStats.R
\name{N_NA}
\alias{N_NA}
\alias{count}
\title{Count of missing (NA and NaN) values}
\usage{
N_NA(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector containing a single value of the count of
  observations that are NA or NaN.
}
\description{
Count the missing values (NA and NaN).
}
\examples{
observations <- c(4, 4.4, 0/0, 1/0, -1/0, NA)
N_NA(observations)

}
\seealso{
Other descriptive functions: 
\code{\link{CI95}()},
\code{\link{CV}()},
\code{\link{L95CI}()},
\code{\link{Max}()},
\code{\link{Mean}()},
\code{\link{Median}()},
\code{\link{Min}()},
\code{\link{N_Obs}()},
\code{\link{N_Zero}()},
\code{\link{SD}()},
\code{\link{SE}()},
\code{\link{Sum}()},
\code{\link{U95CI}()},
\code{\link{Var}()},
\code{\link{sumFunShort}()},
\code{\link{sumFun}()}
}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
\concept{descriptive functions}
