% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeTools.R
\name{now}
\alias{now}
\title{Find the current date and time.}
\usage{
now(...)
}
\value{
The date and time in POSIXct format.
}
\description{
Find the current date and time.
}
\examples{

now()

}
\seealso{
Other time tools: 
\code{\link{minsec2sec}()}
}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
\concept{time tools}
