% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pounds_scrapped}
\alias{pounds_scrapped}
\title{Pounds Scrapped for Department 19}
\format{
\subsection{\code{pounds_scrapped} A data frame with 12 rows and 4 columns:}{

\describe{
\item{date}{Each month for which Pounds Scrapped is reported.}
\item{value}{Number of pounds of scrap per 1000 pounds of finished product.}
}
}
}
\source{
Wheeler, Donald J. Understanding Variation: The Key to Managing
Chaos. 2nd edition, SPC Press, 2000
}
\usage{
pounds_scrapped
}
\description{
Data adapted from Donald J. Wheeler's book Understanding Variation: The Key
to Managing Chaos p76. As originally described on p73: "Department 19
produces a sheet product. The edges of this sheet, and the end of each roll,
have to be trimmed off as part of the production process. This unavoidable
scrap is reported by Department 19 in terms of the number of pounds of scrap
per 1000 pounds of finished product."
}
\keyword{datasets}
