% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick2Sample.R
\name{Quick Two Sample}
\alias{Quick Two Sample}
\alias{q2s}
\alias{quick2sample}
\title{Fast and easy two-sample inferrential tests for summarized data.}
\usage{
q2s(
  est_A = NA,
  est_B = NA,
  SE_A = NA,
  SE_B = NA,
  N_A = as.integer(NA),
  N_B = as.integer(NA),
  CI = 0.95
)
}
\arguments{
\item{est_A}{Numeric vector with length equal 1. The parameter estimate from the first population of interest. This could be a mean or another parameter estimate such as the estimated number of walleye harvested in Lake X.}

\item{est_B}{Numeric vector with length equal 1. The parameter estimate from second population of interest. This could be a mean or another parameter estimate such as the estimated number of walleye harvested in Lake Y.}

\item{SE_A}{Numeric vector with length equal 1. The standard error of the parameter estimate from the first population.}

\item{SE_B}{Numeric vector with length equal 1. The standard error of the parameter estimate from the second population.}

\item{N_A}{Numeric vector with length equal 1. The sample size from the first population. This is optional and only needed to preform the t-test. The t-test is slightly more powerful than the z-test, especially at low sample sizes.}

\item{N_B}{Numeric vector with length equal 1. The sample size from the second population. This is optional and only needed to preform the t-test. The t-test is slightly more powerful than the z-test, especially at low sample sizes.}

\item{CI}{Numeric vector of length = 1 for the width of the confidence interval. Defaults to 0.95.}
}
\value{
A list with the results of the z-test and t-test.
}
\description{
Function runs a simple two-tailed two-sample t-test and z-test on summary data (estimate, standard error, and sample size). This is useful for analyzing summarized data when the raw data is not longer available.  For example, angler creel survey reports typically just include a parameter estimate such as an average catch rate and a standard error.
}
\note{
This is a quick and dirty procedure as a last resort for testing differences when raw data is no longer available but summarized data is. Both the z- and t-test assume that there is a normal distribution of residuals and this is very important for claculating the p-values but cannot be verified when working with summarized data.
  However, the tests presented here are somewhat conservative and that will help prevent Type I errors. The z-test is more conservative than the t-test and this function only preforms two-sided tests which are more conservative (by a factor of 2) then single sided-tests.
}
\examples{
q2s (est_A = 26238, est_B = 77113, SE_A = 3555, SE_B = 13401, N_A = NA, N_B = NA, CI = 0.95)
q2s (26238, 77113, 3555, 13401)

}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
