% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revelio.R
\name{revelio}
\alias{revelio}
\title{Reveal what's in your workspace.}
\usage{
revelio(...)
}
\value{
Technically returns a NULL vector b/c I can't figure out how to force R to not return an object. However, the intent of the function is just print to the screen.
}
\description{
Function prints the libraries currently loaded and all the objects available in the global environment. The objects are organized by their container type.  Vectors are further organized by their data class.
}
\examples{
revelio()

}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
