% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descStats.R
\name{sumFun}
\alias{sumFun}
\title{Collection of descriptive statistics that ignoring missing values}
\usage{
sumFun(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A named numeric vector containing a variety of descriptive statistics that
  ignoring missing values.
}
\description{
Collection of descriptive statistics that ignoring missing
  values.
}
\examples{
observations <- c(0, 1, 2, 4, 5, 6, 7, NA, 9, 10)
sumFun(observations)

}
\seealso{
Other descriptive functions: 
\code{\link{CI95}()},
\code{\link{CV}()},
\code{\link{L95CI}()},
\code{\link{Max}()},
\code{\link{Mean}()},
\code{\link{Median}()},
\code{\link{Min}()},
\code{\link{N_NA}()},
\code{\link{N_Obs}()},
\code{\link{N_Zero}()},
\code{\link{SD}()},
\code{\link{SE}()},
\code{\link{Sum}()},
\code{\link{U95CI}()},
\code{\link{Var}()},
\code{\link{sumFunShort}()}
}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
\concept{descriptive functions}
