% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTools.R
\name{today}
\alias{today}
\title{Find what day is today.}
\usage{
today(...)
}
\value{
The date in R date format.
}
\description{
Find what day is today.
}
\examples{

today()

}
\seealso{
Other date tools: 
\code{\link{ordinalDate}()},
\code{\link{yy2yyyy}()}
}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
\concept{date tools}
