% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmr_stats.R
\name{upper_natural_process_limit}
\alias{upper_natural_process_limit}
\title{Compute the Upper Natural Process Limit}
\usage{
upper_natural_process_limit(avg_moving_range, avg_value)
}
\arguments{
\item{avg_moving_range}{Numeric value, expressing the average of the moving
ranges.}

\item{avg_value}{Numeric value, expressing the average of the values.}
}
\value{
Numeric value.
}
\description{
Compute the Upper Natural Process Limit
}
