% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmr_stats.R
\name{upper_range_limit}
\alias{upper_range_limit}
\title{Compute the Upper Range Limit}
\usage{
upper_range_limit(avg_moving_range)
}
\arguments{
\item{avg_moving_range}{Numeric value, expressing the average of the
moving ranges.}
}
\value{
Numeric value.
}
\description{
Compute the Upper Range Limit
}
