% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/westWing.R
\name{westWing}
\alias{westWing}
\title{Pick random episodes of the West Wing tv show.}
\usage{
westWing(N = 1)
}
\arguments{
\item{N}{Numeric. The number of random episodes to choose.}
}
\value{
A vector of West Wing episodes in 'Season-Episode' format.
}
\description{
Pick random episodes of the West Wing tv show.
}
\examples{

westWing()

westWing(1000)

}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}
}
