% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheelr-package.R
\docType{package}
\name{wheelr-package}
\alias{wheelr}
\alias{wheelr-package}
\title{wheelr: Understanding Variation}
\description{
A simple tool for applying statistical process control techniques for understanding variation, based on the teachings of Donald J. Wheeler.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lddurbin/wheelr}
  \item Report bugs at \url{https://github.com/lddurbin/wheelr/issues}
}

}
\author{
\strong{Maintainer}: Lee Durbin \email{l.d.durbin@gmail.com}

}
\keyword{internal}
