% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheelR.R
\docType{package}
\name{wheelR}
\alias{wheelR}
\title{Powell Wheeler's Miscellanious Tools}
\description{
Powell Wheeler's Miscellanious Tools.
}
\details{
Functions from this package can be used to perform a variety of tasks.
}
