% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z2A.R
\name{Z2A}
\alias{Z2A}
\title{Convert instantaneous mortality (Z) to the annual mortality (A) rate.}
\usage{
Z2A(Z, SE_Z = NA, CI = 0.95)
}
\arguments{
\item{Z}{Vector of length = 1 of an instantaneous mortality rate. This can be entered as a positive or negative number.}

\item{SE_Z}{Vector of length = 1 of Standard Error of the instantaneous mortality rate. Defaults to NA.}

\item{CI}{Vector of length = 1 for the width of the confidence interval. Defaults to 0.95.}
}
\value{
A list with with the estimates for A and Z, their standard errors, and their 95% confidence intervals.
}
\description{
Quickly convert a difficult to interprete instantaneous mortality to a more understandable annual mortality rate. Also, converts the SE(Z) to SE(A) and bootstraps confidence intervals with the gamma distribution.
}
\section{testing}{
  Tested against the commonly known relationship that a Z of 0.693 equals 50% A.  For example page 204 in Ogle, D.H. 2016. Introductory Fisheries Analyses with R. Chapman & Hall/CRC, Boca Raton, FL.
}

\examples{

Z2A(0.69)

Z2A(0.69, 0.1)

Z2A(0.69, 0.1, 0.99)

}
\author{
A. Powell Wheeler, \email{powell.wheeler@gmail.com}

Kyle Rachales
}
