% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheelr-package.R
\docType{package}
\name{wheelr-package}
\alias{wheelr}
\alias{wheelr-package}
\title{wheelr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lddurbin/wheelr}
  \item Report bugs at \url{https://github.com/lddurbin/wheelr/issues}
}

}
\author{
\strong{Maintainer}: Lee Durbin \email{l.d.durbin@gmail.com}

}
\keyword{internal}
